Create ifelsedivisibility.cpp
This C++ program prompts the user to enter an integer and checks if it is divisible by 5. It uses the modulus operator (%) to find the remainder when dividing by 5. If the remainder is zero, the program outputs that the number is divisible by 5; otherwise, it states it is not divisible by 5. The program then ends.