Create divisibleby5and3.cpp
This C++ program checks if a given number is divisible by both 3 and 5. It first prompts the user to enter a number and stores the input. Then it applies the modulus operator to test divisibility. If the number is divisible by both 3 and 5, it prints that it is divisible; otherwise, it prints that it is not divisible.