Create divisibleby5or3.cpp
This C++ program checks if a number entered by the user is divisible by 3 or 5. It first prompts the user for input, then uses the modulus operator to test divisibility. If the number is divisible by either 3 or 5, it prints that the number is divisible; otherwise, it prints that it is not divisible.