Create PrintOddNo.cpp
This program prints all odd numbers from 1 to n. The user inputs a number n. Inside the for loop, each number is checked: if it’s even (i % 2 == 0), the continue statement skips that iteration. Otherwise, it prints the number. The result is a list of odd numbers up to n, each printed on a new line.