Create Countnoofdigit.cpp
This program counts the number of digits in a given number. The user enters a number n. Inside the while loop, the number is divided by 10 each time, which removes the last digit. The variable count increases with each loop until n becomes 0. Finally, the program prints how many digits were in the original number.