Create Swapusingextravariable.cpp
This C program swaps the values of two integer variables using a temporary variable. It first initializes x and y with 5 and 6, displays them, then stores the value of x in t, assigns y to x, and finally assigns t (old x) to y. After swapping, it prints the new values showing that x=6 and y=5.