Create PowerRecursion.cpp
This C++ program computes the power of a number using recursion. The pow() function takes two integers, base a and exponent b. If b is 0, it returns 1; otherwise, it returns a * pow(a, b - 1). The program reads both numbers from the user and prints the result of raising a to the power b.