Create Powerlogarithmic.cpp
This C++ program computes the power of a number efficiently using recursion and the divide-and-conquer method. The power() function calculates a^b. If b is even, it returns (a^(b/2))²; if odd, it multiplies once more by a. The program takes two numbers a and b as input and prints a raised to the power b.